Fix Eject USB code, and add icons to UI
The Eject USB code has been fixed and modified to work with both x86 and
x64 systems, and icons were added to the Open Tray and Eject USB buttons
(still looking for one for Close Tray).

+++ UsbEject/UsbEject/Volume.cs
@@ -36,7 +36,7 @@ namespace BetterExplorer.UsbEject
                 if (_volumeName == null)
                 {
                     StringBuilder sb = new StringBuilder(1024);
-                    if (!Native.GetVolumeNameForVolumeMountPoint(Path + "\\", sb, sb.Capacity))
+                    if (!Native.GetVolumeNameForVolumeMountPoint(Path + "\\", sb, (uint)sb.Capacity))
                     {
                         // throw new Win32Exception(Marshal.GetLastWin32Error());
